MINOR: readfiles: check if output is ELF binary
One of the checks for /proc/self/exe returns a binary directly to
terminal, which could print a several MB file out and also can cause
pain within a terminal due to a variety of escape sequences requiring a
"reset" to be performed.

This commit adds a check for ELF magic bytes to determine whether the
output is an ELF binary and if so it avoid writing it to stdout but
still writes it a local file.

Sample output:
```
[INFO]:Reading file : /proc/self/exe
ELF binary found - not printing to stdout
[INFO]:Writing file : /proc/self/exe to 127.0.0.1/_proc_self_exe
```